peer_rec_sys/requirements.txt
astroid==2.4.2
bcrypt==3.2.0
cachelib==0.1.1
cffi==1.14.3
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.0.1
pycparser==2.20
pylint==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.6
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
